refactor:(IntegrityChecks) more refactor wip
