+++ src/Validators/FormSchema/IntegrityChecks/Behaviours/CurrentEnvironmentSubmitSlugsCheck.cs
@@ -1,8 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
-using form_builder.Enum;
 using form_builder.Extensions;
 using form_builder.Models;
 using Microsoft.AspNetCore.Hosting;
@@ -17,20 +15,24 @@ namespace form_builder.Validators.IntegrityChecks.Behaviours
             _environment= environment;
         }
 
-        public IntegrityCheckResult Validate(Behaviour behaviour)
+        public IntegrityCheckResult Validate(List<Behaviour> behaviours, string formName)
         {
             var integrityCheckResult = new IntegrityCheckResult();
 
             var foundEnvironmentSubmitSlug = false;
-            foreach (var subItem in behaviour.SubmitSlugs.Where(subItem => subItem.Environment.ToLower().Equals(_environment.EnvironmentName.ToS3EnvPrefix().ToLower())))
-                foundEnvironmentSubmitSlug = true;
+
+            foreach (var behaviour in behaviours)
+            {
+                foreach (var subItem in behaviour.SubmitSlugs.Where(subItem => subItem.Environment.ToLower().Equals(_environment.EnvironmentName.ToS3EnvPrefix().ToLower())))
+                    foundEnvironmentSubmitSlug = true;
+            }
 
             if (!foundEnvironmentSubmitSlug)
-                integrityCheckResult.AddFailureMessage($"No SubmitSlug found in form '{schema.FormName}' for environment '{_environment.EnvironmentName}'.");
+                integrityCheckResult.AddFailureMessage($"No SubmitSlug found in form '{formName}' for environment '{_environment.EnvironmentName}'.");
 
             return integrityCheckResult;
         }
 
-        public async Task<IntegrityCheckResult> ValidateAsync(Behaviour behaviour) => await Task.Run(() => Validate(behaviour));
+        public async Task<IntegrityCheckResult> ValidateAsync(List<Behaviour> behaviours, string formName) => await Task.Run(() => Validate(behaviours, formName));
     }
 }
