+++ src/Validators/FormSchema/IntegrityChecks/Behaviours/EmptyBehaviourSlugsCheck.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using form_builder.Models;
 
@@ -5,16 +6,18 @@ namespace form_builder.Validators.IntegrityChecks.Behaviours
 {
     public class EmptyBehaviourSlugsCheck: IBehaviourSchemaIntegrityCheck
     {
-        public IntegrityCheckResult Validate(Behaviour behaviour)
+        public IntegrityCheckResult Validate(List<Behaviour> behaviours, string formName)
         {
             var integrityCheckResult = new IntegrityCheckResult();
 
-            if (string.IsNullOrEmpty(behaviour.PageSlug) && (behaviour.SubmitSlugs == null || behaviour.SubmitSlugs.Count == 0))
-                integrityCheckResult.AddFailureMessage($"Empty Behaviour Slugs Check, Incorrectly configured behaviour slug was discovered in '{schema.FormName}' form");
-
+            foreach (var behaviour in behaviours)
+            {
+                if (string.IsNullOrEmpty(behaviour.PageSlug) && (behaviour.SubmitSlugs == null || behaviour.SubmitSlugs.Count == 0))
+                    integrityCheckResult.AddFailureMessage($"Empty Behaviour Slugs Check, Incorrectly configured behaviour slug was discovered in '{schema.FormName}' form");
+            }
             return integrityCheckResult;
         }
 
-        public async Task<IntegrityCheckResult> ValidateAsync(Behaviour behaviour) => await Task.Run(() => Validate(behaviour));
+        public async Task<IntegrityCheckResult> ValidateAsync(List<Behaviour> behaviours, string formName) => await Task.Run(() => Validate(behaviour, formName));
     }
 }
