+++ src/Validators/FormSchema/IntegrityChecks/Elements/AddressNoManualTextIsSetCheck.cs
@@ -1,30 +1,23 @@
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using form_builder.Enum;
-using form_builder.Models;
+using form_builder.Models.Elements;
+using System.Threading.Tasks;
 
 namespace form_builder.Validators.IntegrityChecks.Elements
 {
-    public class AddressNoManualTextIsSetCheck: IElementSchemaIntegrityCheck
+    public class AddressNoManualTextIsSetCheck : IElementSchemaIntegrityCheck
     {
         public IntegrityCheckResult Validate(IElement element)
         {
-            var integrityCheckResult = new IntegrityCheckResult();
-            var addressElements = schema.Pages.Where(_ => _.Elements != null)
-                .SelectMany(_ => _.Elements)
-                .Where(_ => _.Type == EElementType.Address)
-                .Where(_ => _.Properties.DisableManualAddress)
-                .ToList();
+            IntegrityCheckResult integrityCheckResult = new();
 
-            addressElements.ForEach(element =>
+            if (element.Type.Equals(EElementType.Address) &&
+                element.Properties.DisableManualAddress)
             {
                 if (string.IsNullOrWhiteSpace(element.Properties.NoManualAddressDetailText))
                 {
                     integrityCheckResult.AddFailureMessage($"Address No Manual Text Is Set Check, 'DisableManualAddress' set to 'true', 'NoManualAddressDetailText' must have value in form {schema.FormName}");
                 }
-            });
+            }
 
             return integrityCheckResult;
         }
